Path relink method done.

+++ test/TestDecoders.jl
@@ -6,7 +6,7 @@
 # This code is released under LICENSE.md.
 #
 # Created on:  Apr 20, 2018 by ceandrade
-# Last update: Dec 17, 2018 by ceandrade
+# Last update: Dec 18, 2018 by ceandrade
 #
 # THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 # AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
@@ -28,7 +28,7 @@ function decode!(chromosome::Array{Float64}, instance::Instance,
     if rewrite
         chromosome .= tmp
     end
-    return sum(tmp)
+    return Float64(sum(tmp))
 end
 
 ################################################################################
@@ -47,5 +47,5 @@ function rank_decode!(chromosome::Array{Float64}, instance::Instance,
     if rewrite
         chromosome .= tmp
     end
-    return rank
+    return Float64(rank)
 end
